feat: Add functionality to automatically add users to distribution groups
This commit adds the ability to automatically add users to distribution groups. It introduces the `New-MailContact` cmdlet to create a new mail contact for each email address in the `$emails` array. Then, it uses the `Add-DistributionGroupMember` cmdlet to add the newly created mail contact to the specified distribution group.

This feature will streamline the process of managing distribution groups and ensure that all relevant users are included.